Moving boilerplate survey code into its own component.
Updates from @rralian's review.
