+++ assets/stylesheets/_components.scss
@@ -71,6 +71,7 @@
 @import 'components/like-button/style';
 @import 'components/logged-out-form/style';
 @import 'components/main/style';
+@import 'components/marketing-survey/cancel-purchase-form/style';
 @import 'components/mobile-back-to-sidebar/style';
 @import 'components/notice/style';
 @import 'components/olark-chatbox/style';
