+++ client/components/marketing-survey/cancel-purchase-form/style.scss
@@ -1,0 +1,4 @@
+.dialog__content .cancel-purchase-form__reason-input {
+	margin: 4px 0 0 24px;
+	width: calc( 100% - 24px );
+}
